Parsers: Add a parser for route information
This is somewhat replacing the controller/method/call values.
'group' is more or less the same as 'controller'
'name' is semantically the same as 'call' was before

'method' was only needed because the handling for 'call' wasn't very
well implemented. The definition was left up to controllers themselves
and it was incorrect more often than not.

The idea behind both new values is that the router sets that information,
so controller don't need to worry about it anymore.